requirements.txt
redis
msgpack-python
